+++ graphistry/layout/graph/graphBase.py
@@ -1,3 +1,5 @@
+# DEPRECRATED: Non-vector operators over non-vectorized data
+
 from graphistry.layout.utils import Poset